Add error handling on One Wire bus
Reading a bit was expected to be "flawless". However, there might
be physical disconnections on the bus and they might be not handled.

Current fix is provided for Linux USART driver, where USB dongle
can be disconnected during operations and now driver would not lock.

Other drivers still pending this locking fix.